chore(prettier): add caching to prettier
this commit upgrade stencil to pretter v2.7.0. with v2.7.0, we now have
the ability to cache the results of running the formatter to speed up
subsequent runs.

prettier provides two caching strategies: 'content' and 'metadata'. the
former was chosen over the latter. although it is likely to be slower,
it appears to have more utility than the latter according to the
prettier documentation:

> In general, metadata is faster. However, content is useful for updating
> the timestamp without changing the file content. This can happen, for
> example, during git operations such as git clone, because it does not
> track file modification times.
> - https://prettier.io/docs/en/cli.html#--cache-strategy